Increased recursion limit.
Because some problems to be solved recursively hit the default stack size
of Python on large inputs, we have decided to directly increase the default
limit when importing the package.

Note that the documentation of sys.recursionlimit says:

The highest possible limit is platform-dependent. A user may need to set the limit higher when she has a program that requires deep recursion and a platform that supports a higher limit. This should be done with care, because a too-high limit can lead to a crash.