refactor: Core read methods in class and tuple requests
Two methods have been added to the main class, `read_homogeneous`
and `read_heterogeneous`, that read arrays of homo- and
heterogeneous type, respectively, without any type checking or
other "wrapper" activity.

A top-level method selector has been added to select which
of the tokenizer methods to use. The intent is that `keep_reading`
only has to determine once which method to use, instead of
resolving it (with the same result) at each iteration.